Breaking out math functions.
